refactor: rename protected dashboard to dashboard and update imports
